v0.0.32
Various updates. Fixes, clarifications, and additions to the
documentation and help functions. Addition of a few new options. Added
error checking and catching for various edge cases in the R code
(examples: chromosome plot when zero reads align to the requested
chromosomes, sample-to-sample comparisons plots when provided with only
one sample, etc).